Move v1 REST endpoints to api module to separate runtime
