+++ pom.xml
@@ -117,6 +117,8 @@
     <immuutables.version>2.3.10</immuutables.version>
     <swagger.annotations.version>1.5.5</swagger.annotations.version>
     <jackson.version>2.7.8</jackson.version>
+    <slf4j.api.version>1.7.10</slf4j.api.version>
+    <junit.version>4.12</junit.version>
 
     <!-- maven plugin version -->
     <maven.enforcer.plugin.version>1.3.1</maven.enforcer.plugin.version>
@@ -133,6 +135,7 @@
     <maven.license.plugin.version>3.0</maven.license.plugin.version>
 
     <fabric8.maven.plugin.version>3.2.8</fabric8.maven.plugin.version>
+    <javaee.api.version>7.0</javaee.api.version>
   </properties>
 
   <modules>
@@ -182,6 +185,13 @@
       </dependency>
 
       <dependency>
+        <groupId>javax</groupId>
+        <artifactId>javaee-api</artifactId>
+        <version>${javaee.api.version}</version>
+        <scope>provided</scope>
+      </dependency>
+
+      <dependency>
         <groupId>org.immutables</groupId>
         <artifactId>value</artifactId>
         <version>${immuutables.version}</version>
@@ -191,6 +201,19 @@
         <groupId>io.swagger</groupId>
         <artifactId>swagger-annotations</artifactId>
         <version>${swagger.annotations.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-api</artifactId>
+        <version>${slf4j.api.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>junit</groupId>
+        <artifactId>junit</artifactId>
+        <version>${junit.version}</version>
+        <scope>test</scope>
       </dependency>
     </dependencies>
   </dependencyManagement>
