--- runtime/src/main/java/com/redhat/ipaas/rest/JacksonContextResolver.java
@@ -1,42 +1,0 @@
-/**
- * Copyright (C) 2016 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.redhat.ipaas.rest;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.ext.ContextResolver;
-import javax.ws.rs.ext.Provider;
-
-// Customized {@code ContextResolver} implementation to pass ObjectMapper to use
-@Provider
-@Produces(MediaType.APPLICATION_JSON)
-public class JacksonContextResolver implements ContextResolver<ObjectMapper> {
-    private ObjectMapper objectMapper;
-
-    public JacksonContextResolver() throws Exception {
-        this.objectMapper = new ObjectMapper().
-            registerModule(new Jdk8Module()).
-            setSerializationInclusion(JsonInclude.Include.NON_EMPTY);
-    }
-
-    public ObjectMapper getContext(Class<?> objectType) {
-        return objectMapper;
-    }
-}
